Improve the code

+++ Master_IoT_GroupProject/src/main.cpp
@@ -6,8 +6,8 @@
 
 typedef struct struct_message {
     int id;
-    int humidity;
-    int temp;
+    float humidity;
+    float temp;
 }struct_message; //struct to receive the data. /!\ Should be the same as the slave /!\ 
 
 struct_message myData; //Initialise the data struct to handle the incoming message from slave. 
@@ -16,8 +16,8 @@ void OnDataRecv(const uint8_t * mac, const uint8_t *incomingData, int len) {
   memcpy(&myData, incomingData, sizeof(myData));
   Serial.println("[+] Data received ");
   Serial.printf("ID: %d\n", myData.id);
-  Serial.printf("Humidity: %d\n", myData.humidity);
-  Serial.printf("Temperature: %d\n", myData.temp);
+  Serial.printf("Humidity: %f\n", myData.humidity);
+  Serial.printf("Temperature: %f\n", myData.temp);
 }
 
 void setup() {
